Put univ year fonctions in dedicated util file
