+++ frontend/src/services/date.js
@@ -93,29 +93,6 @@ const formatMonth = m => {
   return '?';
 };
 
-const currentUnivYear = (separator = '/') => {
-  const SEPTEMBER = 8;
-  const currentMonth = new Date().getMonth();
-  const currentYear = new Date().getFullYear();
-  if (currentMonth < SEPTEMBER) {
-    return `${currentYear - 1}${separator}${currentYear}`;
-  }
-  return `${currentYear}${separator}${currentYear + 1}`;
-};
-
-const getUnivYear = (date, separator = '/') => {
-  const SEPTEMBER = 8;
-  const dateUtc = utcDate(date);
-  const currentMonth = dateUtc.getMonth();
-  const currentYear = dateUtc.getFullYear();
-  if (currentMonth < SEPTEMBER) {
-    return `${currentYear - 1}${separator}${currentYear}`;
-  }
-  return `${currentYear}${separator}${currentYear + 1}`;
-};
-
-const endYearOfCurrentUnivYear = () => parseInt(currentUnivYear().split('/')[1], 10);
-
 export {
   formatFrenchDate,
   formatDDMMYYYY,
@@ -124,7 +101,4 @@ export {
   formatMonth,
   shortFrenchMonthNames,
   utcDate,
-  currentUnivYear,
-  getUnivYear,
-  endYearOfCurrentUnivYear,
 };
